Updated CLI to have more indepth instructions per each command e.g. list add --help will show what description of each aspect of the add command. Delete command now asks user for confirmation before deletion. Made a start on front-end

+++ phase1-backend/commands/commands.js
@@ -3,6 +3,7 @@ import { Command } from "commander";
 import Item from "./model.js";
 import seedData from "../seedData/seed.js";
 import { faker } from "@faker-js/faker";
+import mongoose from "mongoose";
 const program = new Command();
 import {
   addListing,
@@ -37,6 +38,15 @@ const questions = [
   },
 ];
 
+const confirmDeletionPrompt = [
+  {
+    type: "confirm",
+    name: "confirmation",
+    message: "Are you sure you want to delete this listing?",
+    default: false,
+  },
+];
+
 program.version("1.0.0").description("CLI for managing listings");
 
 // Add command
@@ -44,6 +54,13 @@ program
   .command("add")
   .alias("a")
   .description("Add a new listing")
+  .option("-t, --title <title>", "Title of the listing")
+  .option("-d, --description <description>", "Description of the listing")
+  .option("-sp, --start_price <start_price>", "Start price of the listing")
+  .option(
+    "-rp, --reserve_price <reserve_price>",
+    "Reserve price of the listing"
+  )
   .action(() => {
     prompt(questions).then((answers) => {
       addListing(answers);
@@ -55,6 +72,10 @@ program
   .command("find <title>")
   .alias("f")
   .description("Find a listing")
+  .option(
+    "-t, --title <title>",
+    "Title of the listing will bring back listing if it exists in this database"
+  )
   .action((title) => {
     findListing(title);
   });
@@ -64,7 +85,8 @@ program
 program
   .command("update <_id>")
   .alias("u")
-  .description("Update a listing")
+  .description("Update a listing by ID")
+  .option("-i, --_id <_id>", "ID of the listing is required to update")
   .action((_id) => {
     prompt(questions).then((answers) => {
       updateListing(_id, answers);
@@ -76,8 +98,19 @@ program
   .command("delete <_id>")
   .alias("d")
   .description("Delete a listing")
+  .option(
+    "-i, --_id <_id>",
+    "ID of the listing to delete, will be asked for confirmation"
+  )
   .action((_id) => {
-    deleteListing(_id);
+    prompt(confirmDeletionPrompt).then((answers) => {
+      if (answers.confirmation) {
+        deleteListing(_id);
+      } else {
+        console.log("Deletion cancelled");
+        mongoose.connection.close();
+      }
+    });
   });
 
 // List command
@@ -85,6 +118,7 @@ program
   .command("list")
   .alias("l")
   .description("List all listings")
+  .option("-a, --all", "List all listings that exist in the database")
   .action(() => {
     listAllListings();
   });
@@ -94,6 +128,7 @@ program
   .command("seed")
   .alias("s")
   .description("Seed the database with fake data")
+  .option("-s, --seed", "Seed the database with fake data")
   .action(seedData);
 
 program.parse(process.argv);
